+++ gateway-api-integrator/src/resource_manager/http_route.py
@@ -69,7 +69,6 @@ class HTTPRouteResourceDefinition(ResourceDefinition):
     paths: list[str]
     filters: list[dict]
     hostname: str | None
-    additional_hostnames: list[str]
 
     def __init__(
         self,
@@ -119,7 +118,7 @@ class HTTPRouteResourceDefinition(ResourceDefinition):
         Returns:
             The listener id.
         """
-        return f"{self.gateway_name}-{self.http_route_type}-listener"
+        return f"{self.gateway_name}-{self.http_route_type}"
 
     @property
     def http_route_resource_name(self) -> str:
@@ -137,7 +136,7 @@ class HTTPRouteResourceDefinition(ResourceDefinition):
         Returns:
             The list of hostnames or an empty list if none are set.
         """
-        return [] if self.hostname is None else [self.hostname, *self.additional_hostnames]
+        return [] if self.hostname is None else [self.hostname]
 
     def http_route_resource_spec(self, namespace: str) -> dict[str, typing.Any]:
         """Generate a Gateway resource from a gateway resource definition.