+++ docs/changelog.md
@@ -7,6 +7,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 Each revision is versioned by the date of the revision.
 
+## 2026-06-11
+
+- Removed support for `gateway-route` v0 library.
+- Added support for `gateway-route` v1 library.
+- Removed resource creation for `gateway-route` relations.
+- Added support for multiple `gateway-route` relations.
+
 ## 2026-06-01
 
 - Added v1 `gateway-route` interface library.